+++ angular_test/test/matchers/throws_in_angular_test.dart
@@ -3,18 +3,10 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:test/test.dart';
-import 'package:angular/src/facade/exceptions.dart';
 import 'package:angular_test/src/matchers.dart';
 
 void main() {
   test('should catch a non-wrapped exception', () {
     expect(() => throw new StateError('Test'), throwsInAngular(isStateError));
-  });
-
-  test('should catch a view-wrapped exception', () {
-    expect(
-      () => throw new WrappedException('', new StateError('Test'), null, null),
-      throwsInAngular(isStateError),
-    );
   });
 }